Move UniverseObjectVisitor into separate TU files
