+++ UI/FleetWnd.cpp
@@ -14,7 +14,6 @@
 #include "../client/human/HumanClientApp.h"
 #include "../universe/Fleet.h"
 #include "../universe/Planet.h"
-#include "../universe/Predicates.h"
 #include "../universe/Ship.h"
 #include "../universe/ShipDesign.h"
 #include "../universe/System.h"
