+++ msvc2017/Common/Common.vcxproj
@@ -232,6 +232,7 @@
     <ClInclude Include="..\..\universe\Species.h" />
     <ClInclude Include="..\..\universe\System.h" />
     <ClInclude Include="..\..\universe\Tech.h" />
+    <ClInclude Include="..\..\universe\UniverseObjectVisitor.h" />
     <ClInclude Include="..\..\universe\Universe.h" />
     <ClInclude Include="..\..\universe\UnlockableItem.h" />
     <ClInclude Include="..\..\universe\Pathfinder.h" />
@@ -312,6 +313,7 @@
     <ClCompile Include="..\..\universe\Species.cpp" />
     <ClCompile Include="..\..\universe\System.cpp" />
     <ClCompile Include="..\..\universe\Tech.cpp" />
+    <ClCompile Include="..\..\universe\UniverseObjectVisitor.cpp" />
     <ClCompile Include="..\..\universe\Universe.cpp" />
     <ClCompile Include="..\..\universe\UnlockableItem.cpp" />
     <ClCompile Include="..\..\universe\Pathfinder.cpp" />
