+++ universe/Field.cpp
@@ -3,7 +3,7 @@
 #include "Enums.h"
 #include "FieldType.h"
 #include "Meter.h"
-#include "Predicates.h"
+#include "UniverseObjectVisitor.h"
 #include "Universe.h"
 #include "../util/AppInterface.h"
 #include "../util/i18n.h"
