+++ universe/Fighter.cpp
@@ -1,7 +1,7 @@
 #include "Fighter.h"
 
 #include "Enums.h"
-#include "Predicates.h"
+#include "UniverseObjectVisitor.h"
 #include "../Empire/EmpireManager.h"
 #include "../util/AppInterface.h"
 #include "../util/Logger.h"
