+++ universe/Planet.cpp
@@ -5,10 +5,10 @@
 #include "Condition.h"
 #include "Enums.h"
 #include "Fleet.h"
-#include "Predicates.h"
 #include "Ship.h"
 #include "Species.h"
 #include "System.h"
+#include "UniverseObjectVisitor.h"
 #include "Universe.h"
 #include "ValueRef.h"
 #include "../Empire/EmpireManager.h"
